refactor: move WSEditorCommandTree into same dir
