+++ src/web/src/views/workspace/WSEditor/WSEditor.tsx
@@ -4,7 +4,7 @@ import { useParams } from "react-router";
 import { TransitionProps } from "@mui/material/transitions";
 import WSEditorSwaggerPicker from "../WSEditorSwaggerPicker";
 import WSEditorToolBar from "../WSEditorToolBar";
-import WSEditorCommandTree, { CommandTreeLeaf, CommandTreeNode } from "../WSEditorCommandTree";
+import WSEditorCommandTree, { CommandTreeLeaf, CommandTreeNode } from "./WSEditorCommandTree";
 import WSEditorCommandGroupContent, { DecodeResponseCommandGroup } from "../WSEditorCommandGroupContent";
 import WSEditorCommandContent, { DecodeResponseCommand } from "../WSEditorCommandContent";
 import WSEditorClientConfigDialog from "../WSEditorClientConfig";
