+++ src/web/src/views/workspace/WSEditor/index.ts
@@ -1,4 +1,6 @@
 export { WSEditor, WSEditor as default } from "./WSEditor";
+export { default as WSEditorCommandTree } from "./WSEditorCommandTree";
+export type { CommandTreeNode, CommandTreeLeaf } from "./WSEditorCommandTree";
 export { default as WSEditorExportDialog } from "./WSEditorExportDialog";
 export { default as WSEditorDeleteDialog } from "./WSEditorDeleteDialog";
 export { default as WSEditorSwaggerReloadDialog } from "./WSEditorSwaggerReloadDialog";
